# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRelay API",
    "description": "Form orchestration backend.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rotated token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or revoked refresh token"
          }
        }
      }
    },
...
    "/setup/initialize": {
      "post": {
        "tags": [
          "setup"
        ],
        "operationId": "initialize",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Initial admin user created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Already initialized"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "list_users",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InitializeResponse": {
        "type": "object",
        "required": [
          "token",
          "refresh_token",
          "user"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
...
      "NewUser": {
        "type": "object",
        "description": "Input shape for creating a user. `role_ids` is optional and defaults to\nan empty set — the route handler reads it and calls into the RBAC service\nin the same transaction after the user row is inserted.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "RoleSummary": {
        "type": "object",
        "description": "Lightweight role reference — enough to render a badge, populate a\ndropdown, or describe the current user's roles in `/auth/me`. Cheap to\nembed in `UserDto` since no permission join is required.",
        "required": [
          "id",
          "name",
          "is_system"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "is_system": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "TokenPair": {
        "type": "object",
        "description": "Result of a successful `/auth/refresh` rotation: a fresh access token plus\nthe rotated refresh secret (the presented one is now revoked).",
        "required": [
          "token",
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "description": "Outbound representation of a user — what callers see in API responses.\n\n`roles` is populated by `service::populate_roles` (or by the RBAC service\ndirectly for single-user lookups). `From<user::Model>` produces an empty\nroles vec; the responsibility to enrich falls to the caller so this type\nstays cheap to construct in places (login, setup) where the frontend will\nrefetch its session shape via `/auth/me` anyway.",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          }
        }
      },
      "UserList": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}